requirements.txt
openai==0.19.0
replicate==0.4.0
ftfy==6.1.1
accelerate==0.15.0
tokenizers==0.13.2
torch==1.13.1
transformers==4.26.0
tensorflow==2.11.0
keras-cv==0.4.1
diffusers==0.10.2
python-dotenv==0.19.2
selenium==4.8.0
PyAutoGUI==0.9.53
webdriver-manager==3.8.5
emoji==2.2.0